+++ RELEASE_NOTES.md
@@ -1,5 +1,16 @@
 # Release notes
 
+### v0.16.0
+
+**DROPS SUPPORT FOR NODE <18**. As of 24th October 2023, Node 20 is the active
+LTS and Node 18 is maintainence LTS; previous versions are no longer supported.
+
+**LOTS OF `any` CHANGED TO `unknown`**. In particular, errors in the event
+emitter payloads are now `unknown` rather than `any`, so you might need to cast.
+
+Adds support for `graphile-config` - configuration can now be read from a
+`graphile.config.ts` (or `.js`, `.cjs`, etc) file.
+
 ### v0.15.1
 
 Fixes issues with graceful worker shutdowns: